Runs

+++ ex_argv.h
@@ -20,17 +20,17 @@ var char	**argv;
 var char	**argv0;
 var char	*args;
 var char	*args0;
-var short	argc;
-var short	argc0;
-var short	morargc;		/* Used with "More files to edit..." */
+var int	argc;
+var int	argc0;
+var int	morargc;		/* Used with "More files to edit..." */
 
 var int	firstln;		/* From +lineno */
 var char	*firstpat;		/* From +/pat	*/
 
 /* Yech... */
 struct	glob {
-	short	argc;			/* Index of current file in argv */
-	short	argc0;			/* Number of arguments in argv */
+	int	argc;			/* Index of current file in argv */
+	int	argc0;			/* Number of arguments in argv */
 	char	*argv[NARGS + 1];	/* WHAT A WASTE! */
 	char	argspac[NCARGS + sizeof (int)];
 };
